+++ nautilus_trader/adapters/interactive_brokers/historic.py
@@ -362,7 +362,7 @@ def parse_historic_trade_ticks(
             instrument_id=instrument.id,
             price=Price(value=tick.price, precision=instrument.price_precision),
             size=Quantity(value=tick.size, precision=instrument.size_precision),
-            aggressor_side=AggressorSide.NONE,
+            aggressor_side=AggressorSide.NO_AGGRESSOR,
             trade_id=generate_trade_id(
                 ts_event=ts_init,
                 price=tick.price,