+++ nautilus_trader/adapters/interactive_brokers/parsing/instruments.py
@@ -21,10 +21,10 @@ from typing import Union
 from ib_insync import Contract
 from ib_insync import ContractDetails
 
-from nautilus_trader.model.c_enums.asset_class import AssetClassParser
 from nautilus_trader.model.currency import Currency
 from nautilus_trader.model.enums import AssetClass
 from nautilus_trader.model.enums import OptionKind
+from nautilus_trader.model.enums import asset_class_from_str
 from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.identifiers import Symbol
 from nautilus_trader.model.identifiers import Venue
@@ -56,7 +56,7 @@ def sec_type_to_asset_class(sec_type: str):
         "CASH": "FX",
         "BOND": "BOND",
     }
-    return AssetClassParser.from_str_py(mapping.get(sec_type, sec_type))
+    return asset_class_from_str(mapping.get(sec_type, sec_type))
 
 
 def parse_instrument(