+++ nautilus_trader/core/rust/model.pyx
@@ -17,3 +17,5 @@
 
 from nautilus_trader.core.rust.model cimport AccountType  # type: ignore
 from nautilus_trader.core.rust.model cimport AggregationSource  # type: ignore
+from nautilus_trader.core.rust.model cimport AggressorSide  # type: ignore
+from nautilus_trader.core.rust.model cimport AssetClass  # type: ignore