+++ nautilus_trader/model/instruments/base.pyx
@@ -21,8 +21,9 @@ import msgspec
 from libc.stdint cimport uint64_t
 
 from nautilus_trader.core.correctness cimport Condition
-from nautilus_trader.model.c_enums.asset_class cimport AssetClass
-from nautilus_trader.model.c_enums.asset_class cimport AssetClassParser
+from nautilus_trader.core.rust.enums cimport AssetClass
+from nautilus_trader.core.rust.enums cimport asset_class_from_str
+from nautilus_trader.core.rust.enums cimport asset_class_to_str
 from nautilus_trader.model.c_enums.asset_type cimport AssetType
 from nautilus_trader.model.c_enums.asset_type cimport AssetTypeParser
 from nautilus_trader.model.currency cimport Currency
@@ -233,7 +234,7 @@ cdef class Instrument(Data):
             f"{type(self).__name__}"
             f"(id={self.id.to_str()}, "
             f"native_symbol={self.native_symbol}, "
-            f"asset_class={AssetClassParser.to_str(self.asset_class)}, "
+            f"asset_class={asset_class_to_str(self.asset_class)}, "
             f"asset_type={AssetTypeParser.to_str(self.asset_type)}, "
             f"quote_currency={self.quote_currency}, "
             f"is_inverse={self.is_inverse}, "
@@ -263,7 +264,7 @@ cdef class Instrument(Data):
         return Instrument(
             instrument_id=InstrumentId.from_str_c(values["id"]),
             native_symbol=Symbol(values["native_symbol"]),
-            asset_class=AssetClassParser.from_str(values["asset_class"]),
+            asset_class=asset_class_from_str(values["asset_class"]),
             asset_type=AssetTypeParser.from_str(values["asset_type"]),
             quote_currency=Currency.from_str_c(values["quote_currency"]),
             is_inverse=values["is_inverse"],
@@ -294,7 +295,7 @@ cdef class Instrument(Data):
             "type": "Instrument",
             "id": obj.id.to_str(),
             "native_symbol": obj.native_symbol.to_str(),
-            "asset_class": AssetClassParser.to_str(obj.asset_class),
+            "asset_class": asset_class_to_str(obj.asset_class),
             "asset_type": AssetTypeParser.to_str(obj.asset_type),
             "quote_currency": obj.quote_currency.code,
             "is_inverse": obj.is_inverse,