+++ tests/unit_tests/backtest/test_backtest_data_wranglers.py
@@ -175,7 +175,7 @@ class TestTradeTickDataWrangler:
         assert len(ticks) == 100
         assert ticks[0].price == Price.from_str("423.760")
         assert ticks[0].size == Quantity.from_str("2.67900")
-        assert ticks[0].aggressor_side == AggressorSide.SELL
+        assert ticks[0].aggressor_side == AggressorSide.SELLER
         assert ticks[0].trade_id == TradeId("148568980")
         assert ticks[0].ts_event == 1597399200223000064
         assert ticks[0].ts_init == 1597399200223000064
@@ -196,7 +196,7 @@ class TestTradeTickDataWrangler:
         assert len(ticks) == 100
         assert ticks[0].price == Price.from_str("423.760")
         assert ticks[0].size == Quantity.from_str("2.67900")
-        assert ticks[0].aggressor_side == AggressorSide.SELL
+        assert ticks[0].aggressor_side == AggressorSide.SELLER
         assert ticks[0].trade_id == TradeId("148568980")
         assert ticks[0].ts_event == 1597399200223000064
         assert ticks[0].ts_init == 1597399200224000564  # <-- delta diff
@@ -355,7 +355,7 @@ class TestTardisTradeDataWrangler:
         assert len(ticks) == 9999
         assert ticks[0].price == Price.from_str("9682.00")
         assert ticks[0].size == Quantity.from_str("0.132000")
-        assert ticks[0].aggressor_side == AggressorSide.BUY
+        assert ticks[0].aggressor_side == AggressorSide.BUYER
         assert ticks[0].trade_id == TradeId("42377944")
         assert ticks[0].ts_event == 1582329602418379008
         assert ticks[0].ts_init == 1582329602418379008